minor adjustments in main() to output structure Division functionality needs altering...

+++ src/main/java/com/mycompany/arithmeticmethods/Operators.java
@@ -21,7 +21,7 @@ public class Operators {
     public static int multiplication(int a, int b){
         return a * b;
     }
-    public static int division(int a, int b){
+    public static float division(int a, int b){
         return a / b;
     }
     public static int remainder(int a, int b){
